CTS: When running VR mode, Alert windows are always created.
When running in VR Mode, a global App Op restriction is in place
for SYSTEM_ALERT_WINDOW. This has higher precedence than package level
retrictions the test puts in place to test denials. The global App Op
restriction causes the test to fail on devices running in VR mode as this
restriction allows the window to be created, but are made non-visible. The test
expects the window creation to fail.

This patch bypasses the window creation check when running in VR mode.

Bug: 68724913
Test: ran  android.server.cts.AlertWindowsTests
Change-Id: Ic242d5b9a0a44ddd596ca1a652f1aaf2f1570f47

Former-commit-id: ff3ac50ffa847d24056f7c175a802e904fba6e6a