Fix link-type warnings on ctsdeviceutillegacy
The library is used by several CTS test cases that are built with SDK.
However, since the library itself hasn't been built with SDK
(LOCAL_SDK_VERSION is missing), this has been causing link-type check
warnings, which will turn into errors soon.

Fixing the warning by adding LOCAL_SDK_VERSION := current to the
library. In doing so, WebViewOnUiThread.savePicture() was also removed
because this is using WebView.savePicture() which isn't SDK API.
Removing it is safe as nobody is using the method.

Bug: 69899800
Test: m -j CtsGraphicsTestCases CtsMediaTestCases CtsTextTestCases CtsUiRenderingTestCases CtsWebViewStartupApp
was successful and didn't show link-type warning.

Change-Id: Ifce2054b3f40d932da90c603e62bfff85ab82064

Former-commit-id: c922a457bb8a9e8c624d43cf28b01ada3e5b4b89