Build cts-core-test-runner with LOCAL_SDK_VERSION := current
The cts-core-test-runner is using only public API. But it is defined
java:platform which can use internal API, since cts-core-test-runner
is not defined sdk_version.
So, LOCAL_SDK_VERSION := current is added in cts-core-test-runner,
because it can be turn into errors when link type check becomes more
strict.
And LOCAL_SDK_VERSION := current is also added in json and vogarexpect
modules, because these modules are used by cts-core-test-runner.

Bug: 69899800
Test: 1. mma -j under cts-core-test-runner dir
2. mma -j under vogar-expect dir
3. mma -j under json dir
4. mma -j under cts dir

Change-Id: I55d5a2f227a199ae15a1c0f7690c75a05a3cda6f

Former-commit-id: 0bdbc92fc5ba5a5238b37e2a1e1f3d3108bb8368