Add all compatible JVMTI tests to CTS
This adds the tests listed below to CTS. This is all compatible
run-tests as of platform/art commit b5708a93.

Note this does not include test 983 which requires some changes to the
build of libdexfile among other things to work.

Note that the following tests have android.jvmti.cts.needs_new_thread
set to true:
  1927
  1928
  1930

Bug: 73252141
Test: cts-tradefed run commandAndExit cts-jvmti
Change-Id: I428d0bf2e74d26bbaceafba4d04ff7a454b2b867

Former-commit-id: af0ddffb71914b0e25d5c7ac604bac02edd6aa7f