Remove testHideSoftKeyboard_shouldHideKeyboardOnRequest test
As discussed on the associated bug, this test is unable to get to
VR keyboard on Daydream devices and it makes an invalid assumption
that a soft keyboard has an associated View in WindowManager.

Also this test was responsible for a bunch of other issues in the
past. So we decieded to remove this test completely and create a new
one for Android P.

Bug: 64480919
Merged-In: If729bc7966ac9f883e17e71e18b08a786d79c09d
Change-Id: I254dbf6265657d5545f7de01ad835949a731265a

Former-commit-id: 112d2e758eecc40682acb91736660d5f1272b3b4