Remove testHideSoftKeyboard_shouldHideKeyboardOnRequest test
As discussed on the associated bug, this test is unable to get to
VR keyboard on Daydream devices and it makes an invalid assumption
that a soft keyboard has an associated View in WindowManager.

Also this test was responsible for a bunch of other issues in the
past. So we decieded to remove this test completely and create a new
one for Android P.

Bug: 64480919
Change-Id: If729bc7966ac9f883e17e71e18b08a786d79c09d

Former-commit-id: 77acca434bdbafea5e8c6197b5f74738a7af8047