cts: Set LOCAL_SDK_VERSION where possible.
This change sets LOCAL_SDK_VERSION for all packages where
this is possible without breaking the build, and
LOCAL_PRIVATE_PLATFORM_APIS := true otherwise.

Setting one of these two will be made required soon, and this
is a change in preparation for that. Not setting LOCAL_SDK_VERSION
makes the app implicitly depend on the bootclasspath, which is
often not required. This change effectively makes depending on
private apis opt-in rather than opt-out.

Test: make relevant packages
Bug: 73535841
Change-Id: I687153389dd4bfab1c83df5d9af83d89bed76fd4
Merged-In: I5afdfc8e1e523768db9065dd3571cb31bc1dfac5

Former-commit-id: 1cb8cb52ed1c892424def3b5e875581bbaf4c3e1