cts: Set LOCAL_SDK_VERSION where possible.
This change sets LOCAL_SDK_VERSION for all packages where
this is possible without breaking the build, and
LOCAL_PRIVATE_PLATFORM_APIS := true otherwise.

Setting one of these two will be made required soon, and this
is a change in preparation for that. Not setting LOCAL_SDK_VERSION
makes the app implicitly depend on the bootclasspath, which is
often not required. This change effectively makes depending on
private apis opt-in rather than opt-out.

Test: make relevant packages
Bug: 73535841
Change-Id: I74e36863af6cbd5feae0af11f10ff0140fd0958a
Merged-In: I9f39682cf50ac44075b5fea0c3265a10749c0e9e

Former-commit-id: c1998d35a3df9503159f18b2dbb5b380a454618c