Key test: Expect Device ID attestation failure
Use the indication provided by the DevicePolicyManager that Device ID
attestation is not implemented to avoid failing the test.

Previously, the test would assume that Device ID attestation is always
available and would fail the test if it wasn't.

This is wrong as Device ID attestation is an optional feature.
Tested on a GCE instance which does not support Device ID
attestation.

Bug: 72642093
Bug: 73448533
Test: cts-tradefed run commandAndExit  cts-dev -s 127.0.0.1:50487 -a x86_64 -m CtsDevicePolicyManagerTestCases -t com.android.cts.devicepolicy.DeviceOwnerTest#testKeyManagement -l DEBUG
Change-Id: I8845872a30650dbb8e75988b0cc910cbd1771121

Former-commit-id: 7cc320b7f504971eb4077b00c7276301504fc0fa