IME CTS can't rely on SOFT_INPUT_STATE_UNSPECIFIED
As described in Bug 73793727, unfortunately the behavior of
SOFT_INPUT_STATE_UNSPECIFIED is something that
  "the system can do whatever it thinks valid"
hence we cannot use it in our CTS, even though most of Android apps
run with this mode because it's the default value.

Bug 77152727 tracks what our compatibility story for
SOFT_INPUT_STATE_UNSPECIFIED looks like, but in order to make our CTS
tests deterministic and predictable the best thing we can do right now
is to apply SOFT_INPUT_STATE_UNCHANGED for our test Activity.

Bug: 77152727
Fixes: 73793727
Test: atest CtsInputMethodTestCases
Change-Id: I258c015e76efbf9df3c10df6e80c4f969247361e

Former-commit-id: 505c7306b9ae5a7e1c7423e6f024db7d1efc259e