Explicitly set LOCAL_CERTIFICATE on CtsCarrierApiTestCases.
This test depends on having a particular certificate, because it is
uses a test SIM's whitelist of app certificates to gain needed
privileges. The default "testkey" certificate can vary depending on
the device being built, so we should hard-code the default AOSP
testkey certificate to ensure the test always works as expected
regardless of build environment.

Bug: 73884956
Test: Ran tests locally with atest
Change-Id: Iad1711a7f2ed6b3ca7a4b9a4624ea9783db4bb16

Former-commit-id: c84c922c4f87e48ac7d185fca92c0c336529e4b4