Enables initialTouchMode in PreferenceActivityFlow CTS Tests
By default, the first preference is focused when initialTouchMode is
false so that devices with a trackball / similar start with an item
focused. Unfortunately when an activity is recreated, this selection
doesn't always persist which can affect before/after screenshot
comparison. Enabling initialTouchMode when creating the ActivityTestRule
enables normal behaviour for devices which rely on touch, not focusing
anything when the activity is launched.

Bug: 77702043
Bug: 62368423
Bug: 67738829
Test: CtsPreference2TestCases
Change-Id: I454fb9fda36cb68e544ab77370b2d8277e07df29

Former-commit-id: dfc8209f2ede18382b582716dd8db4df8c8b6aaa