DO NOT MERGE (M) Test that changing groups prevents auto-granting
If a run time permission of a group is already granted we grant the
other permission of the group automatically when requested.

Hence if an already granted permission changed its group during an
update suddenly permission of a potentially not approved group will
get auto-granted.

This test makes sure that this won't happen anymore.

Test: atest android.permission.cts.PermissionGroupChange
Bug: 72710897
Change-Id: Ib2165d1ae53b80455ebe02e07775853e37a2e339

Former-commit-id: 84f1c8f350d94b82e3e092983e7077dff9765d7e