Launch activity to secondary stack after entering split-screen.
Tests exercising split-screen assume that the primary split-screen
stack will have focus if launching from an existing activity in said
stack. However, if home appears when entering split-screen, the
primary stack will be minimized and not gain focus. Logic around
launch adjacent is affected by the focused stack and tests relying on
the primary stack gaining focus in this scenario will fail.

This changelist addresses this issue by showing another activity in a
secondary split-screen stack. This will unminimize the primary stack
and allow it to gain focus when launching another activity.

Change-Id: I5fad93bc837d287455e173e2ba106924c89112a8
Fixes: 77966740
Test: atest CtsActivityManagerDeviceTestCases:android.server.am.ActivityManagerSplitScreenTests

Former-commit-id: dbc965f23181d064f36ff5cfbae657db92e0ea97