Test VMDebug.allowHiddenApiReflectionFrom
This patch adds tests of VMDebug.allowHiddenApiReflectionFrom() which
allows to exempt a single class from reflection/JNI hidden API checks
when an app is debuggable.

We test that it works by adding a new killswitch test where the
mechanism is exempting DexMemberChecker.

We test that the mechanism cannot be misused by non-debuggable apps by
extending the hidden API blacklist test to attempt to call the method
but expecting for it to fail.

Bug: 64382372
Test: (cd cts/tests/signature ; ./runSignatureTests.sh)
Change-Id: I1dd8496ecb67e992b6281591b3a9d12df4d6bd56

Former-commit-id: ffdbfcc26549f4b8a4cc71a65f2ebe1adc0f0f6d