[Harness Separation] Move common/util and common/host-side/util to new harness dir
* Why keeping them in harness folder?
The majority of the code in these two folders are shared
by test cases and test harness(cts-tradefed). We don't want
test harness to have any dependency on the test case repo/directory.
Therefore we move these code to the harness folder (and later to the
harness repo).

* Why not splitting common/host-side/util?
One alternative is to split build module "compatibility-host-util"
to
  1) "compatibility-host-util-testcase"
    includes java files that are solely used by test cases
  2) "compatibility-host-util",
    include java files that are shared by test cases and harness.
However, this would require us to modify many hosttests mk files to make them
depend on both the two libraries, which is a lot of work.
And it doesn't really solve the problem as testcases still
depend on utils whose code will be maintained in the harness repo.

The other alternative is to split the build module
"compatibility-host-util" to
  1) "compatibility-host-util":
     includes java files that are shared
  2) "compatibility-host-util-harness" where
     includes java files that are soly used by harness.
If we do this, we'll have harness code depend on something
in the test cases repo/folder, which is undesired. And also
2) would include only 4 files, which doesn't worth the splitting.

* What's the tradeoff made here?
The downside of keeping them at the harness side is that
the backward compatibility with test cases are not maintained. If
such changes are cherry-picked back to older branches, people
have to fix test cases on older branches as well.

* An ideal world in the future
Ideally, we would want that no util code is shared
by testcases and test harness. We will find a way to remove the binary dependency.
Then we push the testcases down to its own JVM and then we will be able
to easily make harness be backward compatible with test cases from older
branches.

Bug: 78461075
Bug: 79201478
Test: make cts -j32 WITH_DEXPREOPT_BOOT_IMG_AND_SYSTEM_SERVER_ONLY=true
TARGET_PRODUCT=aosp_arm64
Test: cts-tradefed run cts --compatibility:include-filter
CtsGestureTestCases
Test: cts/run_unit_tests.sh
Test: make gts -j32 WITH_DEXPREOPT_BOOT_IMG_AND_SYSTEM_SERVER_ONLY=true
TARGET_PRODUCT=aosp_arm64

Change-Id: I094635f2a59e2ccd7ad66d3c9d6be9a1093b9674

Former-commit-id: 21e76483ea7d0258a9177212b1540afdf435badb