Fix concurrency issue in Telecom CTS test.
The verifyPhoneStateListenerCallbacksForCall method had the issue that it
would wait on the mCallbackSemaphore; this was problematic as that
semaphore triggers on the FIRST phone state update.  Its really the second
update that we want to wait on.

Added a "waitUntilConditionIsTrue" which will wait until we see the
state update which is desired and then continue verification.

Test: Ran CTS test; modified to force concurrency scenario.
Bug: 79372691
Merged-In: Id5addc5f82bd52361865d04ebb3a4a515085e5f9
Change-Id: Ic5d986cfe3b023f81a7e59d8efd257c9f0d8aa17

Former-commit-id: 6aa86be14844f7c7a77c6d8b9fbebf5097a581b9