CtsInputMethodServiceHostTestCases for Instant App
This CL enables additional test cases to
InputMethodServiceLifecycleTest to make sure IME lifecycle is working
as expected even when an instant app is being focused.

Those new tests are exact mirrors of existing test cases for full mode
apps.  The only difference is that those new tests launches a test
Activity in an instant app mode.

Fixes: 79988635
Test: atest CtsInputMethodServiceHostTestCases
Test: Manually tested as follows:
  1. adb shell su 0 setenforce permissive
  2. make -j cts
  3. cts-tradefed run cts-instant-dev -m CtsInputMethodServiceHostTestCases
Change-Id: I32bb4535ac2a69d6b2bf544b7b4bd772d609e064

Former-commit-id: 71c2891542c91b9fc75238b15d60a2b2a47ae453