Fixes crashes on Watch device.
Clearkey support is not required on Watch devices.

On tests that do not play content: if the device is a watch,
and there is no clearkey plugin, skip the test.

On tests that play content: if the device is a watch,
skip the test because the device may not have the codec.

Also, Clearkey tests should use CLEARKEY_SCHEME_UUID
for clarity even though the clearkey service supports
both UUIDs.

bug: 77228589

Test: Cts tests on device with clearkey support removed

Test: Cts tests on device that includes legacy clearkey plugin

Test: Cts tests on device with full clearkey support

Test: Cts test
  ANDROID_BUILD_TOP= ./android-cts/toolsefed run cts-dev
  --module CtsMediaTestCases -t android.media.cts.MediaDrmClearkeyTest

Test: Native CTS test
  ANDROID_BUILD_TOP= ./android-cts/toolsefed run cts-dev
  --module CtsMediaTestCases
  -t android.media.cts.NativeMediaDrmClearkeyTest

Change-Id: Icaba5630dff34a09547ccda5fd2fdc65117d1eff

Former-commit-id: 712cacf636b330519dceffd7021b968473cb2894