CTS: Fix cts on gsi for statsd
A few tests are failing because gsi does not load the
SysUI config from vendor. This caused AoD to be
available and enabled by default, but screen doze to
be unavailable. This meant that the screen turned on
immediately after it turned off when AoD kicked in.
This fixes that by turning AoD off and restoring the
initial setting on the relevant tests. It also adds
a delay to reduce flakiness in wakelock duration.

Test: verified on marlin, shiner, walleye, gsi on walleye
Bug: 80077839
Change-Id: I93cb34ec60aebb0a7ab8abb6f8668c4ad249d8c6

Former-commit-id: b813a33cbe062a15edbd98606ae4f053586b71e2