Fix test_linker_namespaces for devices having translated 2nd arch
There was a bug in 3fe6ede57050085a63e90cad26778beaefb61412 [formerly 4c09b70131f04a6dce543b28f128576d6346d848] @ cts.

The test always skips loading the webview plat support library via
System.loadLibrary(). This skip is necessary because loading the lib
will trigger JNI_OnLoad of the lib, which causes crash when the webview
apk isn't loaded in the process.

However, the problem was that the skipped attempt is unconditionally
reported as a successful load. As a result, the test thinks that the
result of dlopen() and System.loadLibrary() for the same library differ.

The problem is now fixed by skipping the webview plat support library
in the main test routine so that only the result of dlopen() is used
for the lib.

Bug: 110438097
Test: atest android.jni.cts.JniStaticTest#test_linker_namespaces on
aosp_walleye and gce_x86

Change-Id: Ifbfe0f73c6cd24ed11109ffecf8425da8b3ba70f

Former-commit-id: 53c7a5af839960cc53cd1f4225eb1b08d5ea42fc