Mark installable jars with installable: true
Most jars don't need to be installed on the device.  Instead of
using java_library and java_library_static to distinguish between
them make java_library and java_library_static identical and use
installable: true to identify the few jars that need to be dexed
and installed on the device.

Bug: 110885583
Test: m checkbuild
Change-Id: I1813809eaf12e57615f273036a916a5c8a1f6f7a

Former-commit-id: 9e5396708aecd687b1ec743009e29f238af9138b