Part 2. of porting remaining DAsm files to Smali in vm-tests-tfConvert
In this CL the .d files with a single class per file but with comments
are ported to smali.

The conversion was done with this script:

    https://paste.googleplex.com/6455987980468224

which generates .dex files with DASM then calls baksmali (v2.2.2) to
convert them back to smali. After the script the comments were manually
copied from the original .d files into the smali files.

The script also verifies the conversion by calling smali on the result and
comparing the 'dexdump -d' output of the the reference and converted .dex
files without the offsets.

The script renames processed .d files to .d.done which are retained here
for reference and will be removed in a follow-up.

Also tested locally on a walleye target phone.

Test: cts-tradefed run cts -m vm-tests-tf
Bug: 70348233

Change-Id: I7111b5f773c47becae59c39fe8d6c43038900de4

Former-commit-id: fb17150fda8a2a47200628352c94a589444d7903