Part 3. of porting remaining DAsm files to Smali in vm-tests-t
In this CL the .d files containing multiple classes or interfaces
are ported to smali. The conversion was done with this script:

    https://paste.googleplex.com/5225164102959104

which generates .dex files with DASM then calls baksmali (v2.2.2) to
convert them to smali. Since smali does not support multiple classes
in a single file, baksmali's output is multiple .smali files, named
after the class/interface they contain. The scripts creates
directories named <original-filename-base>.smalis to organize these
files.

The build files BuildDalvikSuite.java and SmaliBuildStep.java have been
modified to handle the *.smalis directiories.

The script also verifies the conversion by calling smali on the result and
comparing the 'dexdump -d' output of the the reference and converted .dex
files without the offsets.

The script renames processed .d files to .d.done which are retained here
for reference and will be removed in a follow-up.

The following 3 files contained comments which have been manually copied
into the generated smali files:

    cts/tools/vm-tests-tf/src/dot/junit/opcodes/instance_of/d/T_instance_of_2.d
    cts/tools/vm-tests-tf/src/dot/junit/opcodes/aput_object/d/T_aput_object_3.d
    cts/tools/vm-tests-tf/src/dot/junit/opcodes/check_cast/d/T_check_cast_2.d

Also tested locally on a walleye target phone.

Test: cts-tradefed run cts -m vm-tests-tf
Bug: 70348233
Change-Id: Ia9effaab661ac4412c2487b165996db6a7e221f1

Former-commit-id: 5f0720640cbcae7b7cb19567d5c25e29567aaafe