Remove testWarpedDng
Bug: 78878033
Bug: 79474588
Bug: 78120086
Test: This is just removing a test

The prior fix to only run this test on devices with 2 gigs of memory was
not enough to fix it. This image requires an extraordinary amount of
memory, which we cannot rely on being available.

The test was originally introduced as part of b/78120086, but that
already has a CTS test, so this one is superfluous.

Change-Id: Ide58f96e2b0047c0a787ed0bab65b8458c6f1a98

Former-commit-id: 5cb02a22a13ecfddcb5f4d47c07cf152b2b43fc8