fix testProcSelf{,Score}OomAdjSane tests
This test uses assertNotSame to compare ints. The ints are autoboxed
into Integers, and comparing integers with '==' as assertNotSame does
checks the identity of the object, not it's value. This means that the
testProcSelfScoreOomAdjSane test can't fail since the identities will
never match. The other test, testProcSelfOomAdjSane should also always
pass, but since it compares Integers that are within the range that are
cached (-128 to 127) by java, the object identity comparison effectively
works as a value comparison.

Switch test to use AndroidJUnitRunner so that it can use
assertNotEquals.

Test: ensure tests can fail if process doesn't have the right
oom_{,score_}adj settings.
Test: atest android.permission.cts.FileSystemPermissionTest
Bug: 111655352

Change-Id: Ib2785e503ee06a9ad8529cbe81b4f637c8d6927b

Former-commit-id: e499684944c5576b15c7107e0e14636b3ac1350d