Fix construction of algorithm list
In CipherTest, the class initializer was not marked static, and thus
appended DES variants to the list of expected algorithms multiple times
(as many as 36). This caused the list to be too long for certain tests
to complete in the necessary timeout. Mark the initializer an static,
and alse ensure the fix does not recur by concatenating two 'final' lists
to a third list, rather than adding to an existing one.

Bug: 111137163
Test: CtsKeystoreTestCases
Change-Id: Id5d861f351029bf330e36d1e7005f116c267770e
(cherry picked from commit a90bdd378c3f372b06115205bf41aa737c7a0c1b)

Former-commit-id: b452a4047343be5f59e8527a814da70ad1871766