Change assertion logic of PrivappPermissionTest
Even if multiple apps violate PrivappPermissionTest, it aborts just
when violation of one app is detected. So it's too inconvenient to
detect all violations at once.

Move the assertion from inside of  the loop for detecting violation to
outside of it in order to detect all violators.

Test: run cts -m CtsPermission2TestCases -t \
android.permission2.cts.PrivappPermissionsTest#\
testPrivappPermissionsEnforcement

Change-Id: Iecd6f8060845a0ea69934dc7cee225989d8a58f3

Former-commit-id: a34c8cf51c14d8f00e60237dfdf8a613ae376f61