StrongBox support for FingerprintBoundKeysTest
This test now functions by first testing that Auth Bound keys work
properly, both on a per operation basis and when
setUserAuthenticationValididtyDurationSeconds is in use.

Next, if a StrongBox secure element is present on the system and the
feature flag is enabled, the test will be rerun to test that the same
use cases are being correctly handled by StrongBox. This will function
as a good sanity test for ValidationToken handling, as
setUserAuthenticationValididtyDurationSeconds uses a ValidationToken to
keep time synched between the system and the secure element.

Bug: 111443699
Test: "make CtsVerifier" and then install and run the
FingerprintBoundKeysTest under Security in the CtsVerifier App

Change-Id: I20dd34e49e1f7d2284e96db463831628b9df9092

Former-commit-id: 7976d226db6de00a015c36d59fb33875d240a556