Increase timeout of CtsLibcoreWycheproofBCTestCases
The test uses generateSafePrimes() which loops and tries until
it gets suitable value, however, it takes an unpredictable
amount of time.
On less powerful devices, the test timeout of 10mins might not be
enough, leading to test timeouts. Bump the timeout for thist test to
3600mins.

bug: 115712670
Change-Id: Ic86b4cb406ed6c02219b56e89059f692ff716d26
Merged-In: Ic86b4cb406ed6c02219b56e89059f692ff716d26
Signed-off-by: Xinghua Yang <xinghua.yang@mediatek.com>

Former-commit-id: 031c47f6776e10c1dd000d0a59287bf7596ff3f8