VpxEncoderTest: Record decoded video within crop rectangle if given
The encoder quality test scheme is to record the decoded video with frame size
"width" and "height" obtained from the format of decoder. However, "width" and
"height" is the frame size for decoder's allocation (coded size), which may be
larger than the size from encoding demand. For such case, "crop rectangle"
should be specified in format which occupies a portion of the video frame.

That is, the video content within "crop rectangle" is what we need to record
for calculating PSNR rather than the whole frame.

This CL has changed to use "crop rectangle" to record video if given.

Bug: 119247958
Test: run android.media.cts.VpxEncoderTest#testEncoderQualityVP9
Change-Id: I4ee191b58e3f4154c88cc6fe3e69f3a1b4b7ebd0

Former-commit-id: 0c97c7d0bcfb0fb92a0f4d5c02fdee83394eff8e