Add internalTransmitApduWithoutP2() for OmapiTest.testP2Value()
The test case exists for checking whether the underlying implementation
correctly chooses 0x00 as the P2 value when Session.openLogicalChannel()
is used without P2 value. Actually the APDU command for the test is a
case 2 command, so this change also adds 0x00 to the end of the command
as the P3/Le value in order to avoid confusiton.

Change-Id: I223eef3190b011599237644f7a340e5820809670

Former-commit-id: 77eea173203d7b8be62309134f15e3dbc80c9671