Reactivate commented out test logic
This remove the commented out test in TelephonyProviderTest for
equivalent logic in a test under MmsProviderTest.

Note the section in the commented out test:

```
assertNull("...", uri2);
if (uri2 == null) {
    return;
}
```

This section effectively made the rest of the test dead code. The test
added under MmsProviderTest reflects this and only tests the logic that
would have been tested.

Other tests around opening the data file and blocking updates to the
_data column will be added in future work.

Test: Manually verified CTS test passes under correct conditions
Bug: 121044880
Change-Id: I69dd3f1e9bf508ba013fca05ac7d70be4251d343

Former-commit-id: 7f5ab069c6f3f79192984b238bee573daf8e7e32