Ensure that thread IDs are not reused
This should make sure that even when a thread is removed that its thread
ID is not reused for a different set of participants. This is to avoid a
potentially harmful issue where an application might hold a reference to
a thread ID that has been reused by the telephony providers in the
meantime. If an application uses that thread ID through certain paths,
it is possible a message might get sent to the wrong thread.

This test was validated by ensuring that it fails with AUTOINCREMENT
turned off on the threads table, though the test is meant to fail for
other cases where thread IDs might be reused as well.

Test: Manually validated the test case
Bug: 122616805
Change-Id: Iaac984a7df89f869e10b19162154137afc558e87

Former-commit-id: 340740006e5fca6cb394065481d5bb244a3af8d7