[precompiled layouts] Add tests ensuring equivalance of precompiled layouts
Precompiled layouts mean layouts can now be inflated directly from the resource
XML, or by using a precompiled layout file. We need to make sure the the layouts
produced by these two modes are indistinguishable. This tests provides some
coverage for that.

The test inflates several layouts both with and without precompilation, and
makes sure that the generated layouts are structurally similar--that is, that
they have the same classes, IDs, and number of children.

This CL provides baseline test coverage that can be expanded as needed.

Bug: 111895153
Test: atest PrecompiledLayoutTest
Change-Id: I32f756cddcd57129086f1b74b3b8e7fa8017ea43

Former-commit-id: b4a637f2299de6b58ff9e1e97f8b73bf8d8d45f6