[RESTRICT AUTOMERGE] Use different criteria to check SIM state for different purposes
isSimCardPresent() is used to check whether to run the Carrier
Priviledge related test, which is supposed to run only for cellula
devices with SIM card inserted.
isSimCardPresent() is used to check whether to run the Carrier Config
related test, which is supposed to run only when the SIM state is
Absent.

Bug: 119957940
Test: run cts -m CarrierConfigManagerTest
Change-Id: I7d8d9b349633641bd9dd13ab84a98b33e5e001f2

Former-commit-id: 6903b604b510c6b399ce96d569c519bf6721f8d5