CTS tests: opt-in to multi_abi, opt-out of Instant App mode.
This CL covers CtsCompilationTestCases and CtsSaxTestCases. Neither
are affected by instant mode / the system server so it makes no sense
to also run these tests in Instant mode; the normal mode is enough.

CtsCompilationTestCases runs the dex->oat compiler on the device, and
CtsSaxTestCases instruments an XML parser backed by a native
implementation, so it makes sense to run both on multiple ABIs
(e.g. 32bit & 64bit ARM); this is also consistent with all of the
CtsLibcore*TestCases modules doing the same.

Change-Id: I90a60678dadecf1b2a6969af4b5b38008831592a
Fixes: 123605061
Fixes: 123604885
Test: atest CtsCompilationTestCases
Test: atest CtsSaxTestCases

Former-commit-id: 38dfb442ca712ba792804835f55a678db9f2533d