Remove workaround for Retention.SOURCE annotations
The jack toolchain would discard Retention.SOURCE annotation classes
from the runtime making it impossible for the signature tests to access
them using reflection. A whitelist of classes to ignore was added to
workaround that issue. After switching to javac toolchain that is no
longer an issue and so the workaround can be removed.

Removed a duplicate method from SignatureMultiLibsTest and simplified
a stream invocation.

Bug: 63383787
Test: atest
Change-Id: I5d21d295a18b89192c6c32e788d19e3188e63fa3

Former-commit-id: 7ca91b27f3db421d5237263c24d383a091b3362a