CTS: Optimize HiddenApiTest for file-based inputs
Allow parallel access to test data by using random access to the
file. Then the checking can be done in parallel.

This is done by mapping the file to a byte buffer, and scanning for
lines by newline bytes. An iterator can split by a search for a
newline character from the middle of the remaining input sequence.

This is only valid for inputs that do not use unicode characters
violating the simple scan. This is currently given, but actual
testing is future work.

Bug: 123986482
Test: atest --test-mapping cts/tests/signature
Change-Id: Ia3e664e40f467c1bc1b6bf86e682c74a49da1536

Former-commit-id: 3bbd2b1c0dde00579a33e228d74ef8aa56a0589a