Count actual allocated blocks considering inlined data
If data is inlined in inode, aka inline_data in ext4/f2fs, stat() gives
512 bytes st_blocks virtually in order for user to be aware of being blocks
allocated. But, those were not actually allocated and recorded into quota.

Fortunately, we have st_blksize which represents block allocation size in
filesystem, so if st_blocks is not aligned to st_blksize, we can assume
it was allocated virtually.

The final solution is to expose i_blocks directly to user, but currently
there is no POSIX interface for that. We can do that by dump.f2fs which would
be a quite ugly approach.

Bug: 123552461
Change-Id: Iae81f8f82a54b8597144653c22c24dcc068c804c
Signed-off-by: Jaegeuk Kim <jaegeuk@google.com>

Former-commit-id: 4a51d2961f3cd79632ef5a59d74bbae8424c88d7