CTS: Add test for run_as process connecting to app's socket
This test is to verify the policy
"allow runas_app untrusted_app_all:unix_stream_socket connectto"
https://android.googlesource.com/platform/system/sepolicy/+/bafea958fb953f052627793bb50e903db4029769

The test first launches a Java app ('DomainSocketActivity')
that opens a domain socket, being ready to send out a string
when a client connects to it.

Then the test starts a run_as process. The run_as process
executes a native program ('connector') that attempts
to connect to the socket opened by the app and print
out the message it receives from the server.

Bug: 120445954
Test: atest -bit RunasPermissionsTest
The added test passes on a build including the
policy change, and fails on a build prior to the
policy change.

Change-Id: I70efe94eb65e6b278962520008a8121db3a90465

Former-commit-id: 57f5351803335f1d69f11f54cf57104494e9a9c4