Disable isolated storage for tests using DynamicConfigDeviceSide
A recent change to enable isolated storage by default
(Id1b0d40e628d9d790143d50e7c6f5fa06ebd8b38) broke these tests as they
rely on the host side being able to store a file in
/sdcard/dynamic-config-files/ that the app on the device can then read.
Unfortunately, if isolated storage is enabled then the app sees a
different /sdcard/ directory to the one in which the host pushed the
file and so it fails.

The workaround is to disable isolated storage for these tests until the
DynamicConfig mechanism can be updated to work with isolated storage.

Bug: 126565285
Test: atest -p cts/tests/signature
Change-Id: Ie4f2bbe20f0a809e3b7850eabf24f66fa9b6a940
Merged-In: I5880402b0de8a80885d243af50f752ff732c2f38

Former-commit-id: c7a5d1ff03ba57f0f8246822fc7fe7cfb4d32057