Test routing changed callback when first audio device is selected.
As routed device may not be the same as the preferred device, it may
cause routing changed callback test failure if it check the routed
device is the same as preferred device. Instead, it is more robust to
test if the callback is triggered or not when there is routing changed
event.

Bug: 126787597
Test: run RoutingTest
Merged-In: I8b247d91e49b2c099c2cdd06ea0a99ddd37daa03
Change-Id: I8b247d91e49b2c099c2cdd06ea0a99ddd37daa03

Former-commit-id: dd38fee00f5ff792e15ecefc3f4e772f7413c56f